+++ tests/variables.global.test.mjs
@@ -37,7 +37,7 @@ test('variables.global', () => {
     expect(is_contract).not.toBeUndefined();
     expect(is_iterable).not.toBeUndefined();
     expect(is_countable).not.toBeUndefined();
-    expect(objval).not.toBeUndefined();
+    expect(to_object).not.toBeUndefined();
     expect(unserialize).not.toBeUndefined();
     expect(var_export).not.toBeUndefined();
 });