+++ CHANGELOG.md
@@ -33,6 +33,7 @@
     -   [to_double](./docs/variables/to_double.md) &mdash; `doubleval`
     -   [to_float](./docs/variables/to_float.md) &mdash; `floatval`
     -   [to_int](./docs/variables/to_int.md) &mdash; `intval`
+    -   [to_object](./docs/variables/to_object.md) &mdash; `objval`
 
 -   Добавлено в модуль [variables](./docs/variables.md):